change {admisc}	R Documentation

Generic function to change the structure of an object, function of the (changed) parameters used to create it.

Description

A generic function that applies different altering methods for different types of objects (of certain classes).

Usage

change(x, ...)

Arguments

x
An object of a particular class.

...
Arguments to be passed to a specific method.

Details

For the time being, this function is designed to change truth table objects (only). Future versions will likely add class methods for different other objects.

Value

The changed object.

Author(s)

Adrian Dusa

Examples

## Not run: 
# An example to change a QCA truth table
library(QCA)

ttLF <- truthTable(LF, outcome = SURV, incl.cut = 0.8)
minimize(ttLF, include = "?")

# excluding contradictory simplifying assumptions
minimize(
    change(ttLF, exclude = findRows(type = 2)),
    include = "?"
)

## End(Not run)

[Package admisc version 0.39 Index]