Number equality {admisc}	R Documentation

Check difference and / or (in)equality of numbers

Description

Check if one number is greater / lower than (or equal to) another.

Usage

agtb(a, b, bincat)
altb(a, b, bincat)
agteb(a, b, bincat)
alteb(a, b, bincat)
aeqb(a, b, bincat)
aneqb(a, b, bincat)

Arguments

a
Numerical vector

b
Numerical vector

bincat
Binary categorization values, an atomic vector of length 2

Details

Not all numbers (especially the decimal ones) can be represented exactly in floating point arithmetic, and their arithmetic may not give the normal expected result.

This set of functions check for the in(equality) between two numerical vectors a and b, with the following name convention:

gt means “greater than”

lt means a “lower than” b

gte means a “greater than or equal to” b

lte means a “lower than or equal to” b

eq means a “equal to” b

neq means a “not equal to” b

The argument values is useful to replace the TRUE / FALSE values with custom categories.

Author(s)

Adrian Dusa

References

Goldberg, David (1991) "What Every Computer Scientist Should Know About Floating-point Arithmetic", ACM Computing Surveys vol.23, no.1, pp.5-48, doi: 10.1145/103162.103163

[Package admisc version 0.39 Index]