.rda functions: listRDA, objRDA {admisc}	R Documentation

Load and list objects from an .rda file

Description

Utility functions to read the names and load the objects from an .rda file, into an R list.

Usage

listRDA(.filename)

objRDA(.filename)

Arguments

.filename
The path to the file where the R object is saved.

Details

Files with the extension .rda are routinely created using the base function save().

The function listRDA() loads the object(s) from the .rda file into a list, preserving the object names in the list components.

The .rda file can naturally be loaded with the base load() function, but in doing so the containing objects will overwrite any existing objects with the same names.

The function objRDA() returns the names of the objects from the .rda file.

Value

A list, containing the objects from the loaded .rda file.

Author(s)

Adrian Dusa

[Package admisc version 0.39 Index]