Tilde operations {admisc}	R Documentation

Tilde operations

Description

Checks and changes expressions containing set negations using a tilde.

Usage

hastilde(x)
notilde(x)
tilde1st(x)

Arguments

x
A vector of values

Details

Boolean expressions can be negated in various ways. For binary crisp and fuzzy sets, one of the most straightforward ways to invert the set membership scores is to subtract them from 1. This is both possible using R vectors and also often used to signal a negation in SOP (sum of products) expressions.

Some other times, SOP expressions can signal a set negation (also known as the absence of a causal condition) by using lower case letters, while upper case letters are used to signal the presence of a causal condition. SOP expressions also use a tilde to signal a set negation, immediately preceding the set name.

This set of functions detect when and if a set present in a SOP expression contains a tilde (function hastilde), whether the entire expression begins with a tilde (function tilde1st).

Author(s)

Adrian Dusa

Examples

hastilde("~A")

[Package admisc version 0.39 Index]